requirements.txt
Flask==0.12
Flask-Admin==1.5.0
Flask-API==0.7.1
flask-jsontools==0.1.1.post0
flask-login
Flask-MySQL==1.4.0
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.2
Flask-Security>=1.7.5
